Refactor frontend exposing code in API
