--- api/obs/api/routes/__init__.py
@@ -1,8 +1,0 @@
-from . import (
-    info,
-    login,
-    stats,
-    tiles,
-    tracks,
-    users,
-)
